# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Syrian Car Market API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email or phone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /ai/generate-description:
    post:
      operationId: generateCarDescription
      tags: [ai]
      summary: Generate car listing description using AI
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateDescriptionBody"
      responses:
        "200":
          description: Generated description
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedDescription"
  /ai/estimate-price:
    post:
      operationId: estimatePrice
      tags: [ai]
      summary: Estimate car price using AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EstimatePriceBody"
      responses:
        "200":
          description: Price estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceEstimate"
  /ai/recommendations:
    get:
      operationId: getRecommendations
      tags: [ai]
      summary: Get personalized car recommendations
      security:
        - bearerAuth: []
      parameters:
        - name: province
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Recommended cars
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Car"
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /admin/users/{id}:
    patch:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Update user (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    delete:
      operationId: adminDeleteUser
      tags: [admin]
      summary: Delete user (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: User deleted
  /admin/cars:
    get:
      operationId: adminListCars
      tags: [admin]
      summary: List all cars (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All cars
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarListResponse"
  /admin/cars/{id}:
    delete:
      operationId: adminDeleteCar
      tags: [admin]
      summary: Delete a car listing (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Car deleted
  /admin/settings:
    get:
      operationId: getSettings
      tags: [admin]
      summary: Get platform settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSettings"

    patch:
      operationId: updateSettings
      tags: [admin]
      summary: Update platform settings (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsBody"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSettings"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        password:
          type: string
        role:
          type: string
          enum: [buyer, seller]
      required:
        - name
        - password
    LoginBody:
      type: object
      properties:
        identifier:
          type: string
        password:
          type: string
      required:
        - identifier
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        role:
          type: string
        profilePhoto:
          type: ["string", "null"]
        province:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        isVerified:
          type: boolean
        isPremium:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - role
        - isVerified
        - isPremium
        - createdAt
...
    AdminUpdateUserBody:
      type: object
      properties:
        role:
          type: string
        isVerified:
          type: boolean
        isPremium:
          type: boolean
        isBanned:
          type: boolean
    Car:
      type: object
      properties:
        id:
          type: integer
        sellerId:
          type: integer
        sellerName:
          type: string
        sellerPhoto:
          type: ["string", "null"]
        brand:
          type: string
        model:
          type: string
        year:
          type: integer
        price:
          type: number
        mileage:
          type: integer
        fuelType:
          type: string
        transmission:
          type: string
        province:
          type: string
        city:
          type: string
        saleType:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        isHighlighted:
          type: boolean
        isActive:
          type: boolean
        primaryImage:
          type: ["string", "null"]
        viewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sellerId
        - sellerName
        - brand
        - model
        - year
        - price
        - mileage
        - fuelType
        - transmission
        - province
        - city
        - saleType
        - category
        - isFeatured
        - isHighlighted
        - isActive
        - viewCount
        - createdAt
...
    CarListResponse:
      type: object
      properties:
        cars:
          type: array
          items:
            $ref: "#/components/schemas/Car"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - cars
        - total
        - page
        - limit
...
    GenerateDescriptionBody:
      type: object
      properties:
        brand:
          type: string
        model:
          type: string
        year:
          type: integer
        mileage:
          type: integer
        fuelType:
          type: string
        transmission:
          type: string
        condition:
          type: string
        additionalNotes:
          type: string
      required:
        - brand
        - model
        - year
    GeneratedDescription:
      type: object
      properties:
        description:
          type: string
      required:
        - description
    EstimatePriceBody:
      type: object
      properties:
        brand:
          type: string
        model:
          type: string
        year:
          type: integer
        mileage:
          type: integer
        condition:
          type: string
        fuelType:
          type: string
        transmission:
          type: string
      required:
        - brand
        - model
        - year
    PriceEstimate:
      type: object
      properties:
        minPrice:
          type: number
        maxPrice:
          type: number
        suggestedPrice:
          type: number
        currency:
          type: string
        reasoning:
          type: string
      required:
        - minPrice
        - maxPrice
        - suggestedPrice
        - currency
        - reasoning
    PlatformSettings:
      type: object
      properties:
        id:
          type: integer
        primaryColor:
          type: string
        secondaryColor:
          type: string
        logoUrl:
          type: ["string", "null"]
        heroImageUrl:
          type: ["string", "null"]
        backgroundImageUrl:
          type: ["string", "null"]
        siteName:
          type: string
        allowRegistrations:
          type: boolean
        maxListingsPerUser:
          type: integer
        featuredListingPrice:
          type: number
        premiumSubscriptionPrice:
          type: number
      required:
        - id
        - primaryColor
        - secondaryColor
        - siteName
        - allowRegistrations
        - maxListingsPerUser
        - featuredListingPrice
        - premiumSubscriptionPrice
    UpdateSettingsBody:
      type: object
      properties:
        primaryColor:
          type: string
        secondaryColor:
          type: string
        logoUrl:
          type: string
        heroImageUrl:
          type: string
        backgroundImageUrl:
          type: string
        siteName:
          type: string
        allowRegistrations:
          type: boolean
        maxListingsPerUser:
          type: integer
        featuredListingPrice:
          type: number
        premiumSubscriptionPrice:
          type: number
...