# lib/api-zod/src/generated/types/priceEstimate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Syrian Car Market API
 * OpenAPI spec version: 0.1.0
 */

export interface PriceEstimate {
  minPrice: number;
  maxPrice: number;
  suggestedPrice: number;
  currency: string;
  reasoning: string;
}