# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Syrian Car Market API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  role: string;
  /** @nullable */
  profilePhoto?: string | null;
  /** @nullable */
  province?: string | null;
  /** @nullable */
  city?: string | null;
  isVerified: boolean;
  isPremium: boolean;
  createdAt: Date;
}